chore: code clean up and separating modal from poke details
