+++ src/components/modal.tsx
@@ -1,0 +1,23 @@
+interface ModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  children: React.ReactNode;
+}
+
+export function Modal({ isOpen, onClose, children }: ModalProps) {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50 p-4">
+      <div className="relative h-[600px] w-full max-w-2xl rounded-lg bg-white p-6 shadow-xl">
+        <button
+          onClick={onClose}
+          className="absolute right-4 top-4 text-gray-500 hover:text-gray-700"
+        >
+          ✕
+        </button>
+        <div className="h-full overflow-y-auto">{children}</div>
+      </div>
+    </div>
+  );
+}
